+++ djangojs/conf.py
@@ -9,6 +9,8 @@ from django.dispatch import receiver
 DEFAULTS = {
     'DEBUG': True,
     'TESTING': 'test' in sys.argv,
+    'JS_URLS_NAMESPACES': None,
+    'JS_URLS_NAMESPACES_EXCLUDE': None,
 }
 
 # All required settings keys to import rom project